Added if clause to ensure the table does exist for dropping if needed

+++ db/migrate/20250813232513_drop_prompts_for_trainings_table.rb
@@ -1,5 +1,9 @@
 class DropPromptsForTrainingsTable < ActiveRecord::Migration[8.0]
   def change
-	drop_table :prompt_for_trainings
+	if table_exists?(:prompt_for_trainings)
+		drop_table :prompt_for_trainings
+    else
+		puts "Table prompt_for_trainings does not exist"
+    end
   end
 end
